refactor(mcp): relocate test-only accessors to mcp_server_testonly.hpp (#2385)
Move ToolSecurityRow and the three tool_*_for_test() accessors out of the
production mcp_server.hpp into a dedicated mcp_server_testonly.hpp, so the
public header carries no test-only surface. Definitions stay in
mcp_server.cpp (now includes the new header); the [mcp][2g] annotation
cross-check test includes it alongside mcp_server.hpp. No functional change.
