+++ tests/unit/server/test_mcp_server.cpp
@@ -551,6 +551,7 @@ TEST_CASE("MCP AuditStore: query with mcp_tool field", "[mcp][audit]") {
 // ═══════════════════════════════════════════════════════════════════════════
 
 #include "mcp_server.hpp"
+#include "mcp_server_testonly.hpp"      // tool_*_for_test() accessors (issue #2385)
 
 #include "pg/pg_exec.hpp"               // exec_params — degrade the store in the [pg] degrade test
 #include "pg/pg_pool.hpp"               // PgPool for the query_installed_software [pg] test
